+++ src/GearboyCore.cpp
@@ -120,7 +120,7 @@ void GearboyCore::RunToVBlank(GB_Color* pFrameBuffer)
         bool vblank = false;
         while (!vblank)
         {
-            u8 clockCycles = m_pProcessor->Tick();
+            unsigned int clockCycles = m_pProcessor->Tick();
             vblank = m_pVideo->Tick(clockCycles, pFrameBuffer);
             m_pAudio->Tick(clockCycles);
             m_pInput->Tick(clockCycles);
